+++ lib/module/GitLab.rb
@@ -2,9 +2,6 @@ module Rondabot
   class GitLab < SourceControl
     def initialize params
       super(params)
-
-      # get the user credentials
-      # user_credentials = get_user_credentials(params)
     end
 
     def repository_uri