+++ lib/module/SourceControl.rb
@@ -7,17 +7,22 @@ module Rondabot
     def initialize params
       @credentials = []
       @provider = params[:provider]
+      @repository = params[:repository]
 
-      # A GitHub access token with read access to public repos
-      if params[:github_public_token].nil?
-        raise ArgumentError.new("'github_public_token' param is missing!")
+      # When the repository's visibility is public,
+      # the `github_token` must be an access token with read access to the public repositories.
+      #
+      # When repository visibility is private,
+      # the `github_token` must be an access token with full control of private repositories.
+      if params[:github_token].nil?
+        raise ArgumentError.new("'github_token' param is missing!")
       end
 
       @credentials << {
         "type" => "git_source",
         "host" => "github.com",
         "username" => "x-access-token",
-        "password" => "#{params[:github_public_token]}"
+        "password" => "#{params[:github_token]}"
       }
     end
 
@@ -43,23 +48,7 @@ module Rondabot
         label_language: true
       )
 
-      pull_request = pr_creator.create
-
-      if pull_request&.status == 201
-        content = JSON[pull_request.body]
-        puts "  PR ##{content["pullRequestId"]} submitted"
-      else
-        puts "  PR already exists or an error has occurred"
-      end
-    end
-
-    private
-    def get_user_credentials(params)
-      _credentials = params[:credentials]
-      if _credentials.nil? || _credentials[:username].nil? || _credentials[:password].nil?
-        raise ArgumentError.new("'credentials' param is missing! Check your credentials parameter")
-      end
-      return _credentials
+      return pr_creator.create
     end
   end
 end