+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [1.0.2] - 2020-08-16
+
+- Added support for public and private github repositories
+- Replacing the key pair `username` and `password` with the `acess_token` parameters
+
+## [1.0.1] - 2020-08-14
+
+- Bug fixes and improvements
+
 ## [1.0.0] - 2020-08-14
 
 - Create Rondabot gem dependency